LayoutTests/media/media-source/media-source-changetype-support-expected.txt


EXPECTED (source.readyState == 'closed') OK
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
EXPECTED (sourceBuffer.changeType != 'undefined') OK
END OF TEST
